io: implement multi-threaded io_uring event loop
Implemented a multi-threaded io_uring event loop over the basic async
tcp server that was previously driven by a single-threaded io_uring
event loop for Rheia.

To do this, instead of having a single main thread keeping one file
descriptor, we now spawn a pool of threads where each thread has its own
io_uring file descriptor.

When an inbound connection is established to the tcp listener that we
initialized on the main thread, a coroutine is spawned to handle
incoming messages from the inbound connection. The thread that is picked
for the coroutine to be spawned and driven within is selected via. a
simple round-robin strategy.

Prior to the inbound connection being established though, all threads
except for the main thread will be blocked waiting for the submission or
completion of i/o events to/from io_uring. Therefore, the thread that is
picked for the coroutine to be spawned and driven within needs to be
notified to be unblocked (waken up).

Notifications across threads are done using an eventfd. An eventfd is
assigned to every thread, and initialized next to the code that
initializes the io_uring file descriptor for each and every thread. The
moment the eventfd is initialized, a request to read from the eventfd is
made.

Waking up a thread blocked on a syscall awaiting for the submission or
completion of i/o events to/from io_uring is as simple as writing to
their corresponding eventfd's then.

Apart from that, the README.md for Rheia has been updated to include a
brief on my current thoughts of how the architecture of Rheia should
look like.