runtime: support spsc message-passing and single-threaded execution
Implemented @dvyukov's unbounded single-producer single-consumer (SPSC)
queue. All workers upon runtime initialization now instantiate N SPSC
task queues where N is the number of workers that was specified to be
instantiated on startup.

Wrote unit tests which assert that the SPSC queue respects FIFO order
and exemplifies sane behavior given a producer thread and consumer
thread.

A new method Runtime.runOnWorker(A, B, closure) was introduced which
allows for thread A to delegate the execution of a task encased in a
closure to thread B. With Runtime.runOnWorker(), threads may now
schedule work to be done on other threads via. SPSC message passing.

The event loop local to each I/O worker now polls for pending tasks
coming in from other threads on every loop iteration. If any pending
tasks are available, at most 128 pending tasks from each individual SPSC
task queue are popped and executed in FIFO order prior to the event loop
processing any I/O requests, submissions, and completions from io_uring
on every loop iteration.

Workers are now assigned an integer ID from 0..N where N is the total
number of workers that was specified to be instantiated on startup.

The runtime was refactored to now support being initialized with any
number of worker threads, including zero (single-threaded mode).

All shared state has been migrated over to the main thread. All uses of
mutices have been swapped out with message passing which has allowed for
the simplification of several parts of the code base (i.e. Server).

The readme was updated to note that all shared state is now kept on the
main thread (shared-nothing architecture), and that eventfd is used for
cross-thread notifications.

There is a possibility that Rheia may move towards making use of MPSC
queues rather than SPSC queues in the future, given that unbounded MPSC
queues allow for allocations to be intrusive rather than be on-demand.
Steering the message passing API design away from there being any need
for error handling (i.e. b/c of allocation failures) was a decision that
was made for the sake of ensuring that all concurrency-related code in
the code base is sane and readable.