client: gracefully shutdown writer when reader is closed
Intrusively allocate all state required for queuing up writers that are
waiting for more messages to be queued into Client.

Renamed unparkConnectAttempt() to unparkAttemptToConnect() to keep
naming consistent.

Added removeWriter(*Runtime, *Client.Connection) which de-queues a
connection instance that is suspended waiting for more messages to be
queued into Client. If the connection instance turns out to not have
been suspended and queued in the first place, the function will
no-op.

The write loop of a Client.Connection is now signaled to terminate
should the connection no longer have anymore messages queued to be
written to a target destination, should the connection have its read
side closed, or should the client have signaled for a graceful shutdown.
This solves an edge-case scenario where the connections of a Client
whose remote-end was closed would not clean themselves up.

THe function wakeUpWriters() has been removed now that a write loop is
capable of being signaled to terminate simply by closing the read side
of the write loops' underlying connection instance's socket.