net, worker: send/recv packets, client & shutdown livelock fixes
There was an invariant broken in Client where writers that were
delegated messages to send would be erroneously woken up twice. The
problem is that when a writer is dequeued, should new messages be queued
into a Client, the writers' queue entry does not get reset. This is due
to std.TailQueue.remove() and std.TailQueue.popFirst() not resetting the
next/prev pointers of the node that it is passed in. The fix was to
manually reset the next/prev node pointers for every writer that is
popped from the queue of available writers.

There was an edge case when gracefully shutting down Rheia where one of
the worker threads would not terminate. This was due to the conditions
for determining whether or not a Worker cleans itself up being
contingent on whether or not, within the iterations of its work loop,
the number of scheduled incoming tasks that were processed is 0. These
conditions have now been relaxed which fixes this livelock issue upon
graceful shutdown.

Client.write() is now modified to directly take in a
std.SinglyLinkedList([]const u8).Node pointer such that the caller is
responsible for allocating the memory of messages queued into a Client.

Yield points in Client.runIoLoops() were modified to occur after an
'await' is performed. It can be assumed that Client.runIoLoops() and
Client.serve() solely execute on the main thread (thread 0), while other
coroutines such as Client.runReadLoop() and Client.runWriteLoop() solely
execute on the worker thread assigned to a client connection upon
instantiation.

Client and Server read loops now properly take in an unsigned 32-bit
little endian integer length-prefixed frame by buffering data
which is read from a socket into a linear FIFO ring buffer, and parsing
its contents. Using a ring buffer removes the need to perform large
memcpy's which noticeably improves client/server packet parsing
throughput. A single frame corresponds to a single packet. The maximum
size of a frame is enforced to be 65536 bytes. In the case messages that
are larger than 65536 bytes wish to be sent from one peer to another
peer, a higher-level framing protocol should be executed by both
parties.

Loop.poll() now returns the total number of resumed coroutines, rather
than solely the total number of resumed coroutines corresponding to
completed I/O requests.

Server.serveConnection previously did not yield itself to be executed on
its assigned wroker thread. That is now fixed in this commit.

Logs that detail the address a server is bound to are now printed from
Server rather than in Rheia's main entry point.

The main entrypoint is updated to now allocate, construct, and
spam-write Packet's rather than arbitrary bytes.