rheia: rpc, backpressure, lock-free stack, yield & eventfd improvements
A FIFO ring buffer data structure was implemented for keeping track of
pending requests and handling incoming responses across peers. For the
time being, the data structure is locked under a std.Thread.Mutex. The
hope is to either have the data structure be lock-free, or to have the
data structure locked under an async mutex that takes advantage of the
message passing primitives in Rheia. The FIFO ring buffer in this commit
is now instantiated with a capacity of 65,536 entries for every Client
instance.

There is still pending work left to be done for RPC to be fully
implemented: a Client.request() method needs to be done.

Backpressure for queuing messages to be written into a Client/Server
instance has been implemented. The backpressure mechanism takes into
consideration the total amount of heap memory that was allocated and is
being kept by a Client/Server instance. Right now, hard limits are
configured for a Client to enact backpressure should more than 4MiB of
data be queued into a Client instance, and for a Server to enact
backpressure should more than 1MiB of data be queued into a single
Server.Connection instance.

The way read loops for both Client and Server now allocate buffers for
data to be copied into by a recv() call has been modified s.t. for every
recv() call, a 65,536-length byte slice is allocated. By doing this,
benchmarks unveil that smaller allocations in the ring buffers kept in
the read loops of both Client and Server no longer occur which
noticeably improves I/O throughput.

An effort was made to reduce the amount of write() calls that are made
to eventfd across threads by allowing only one write() to occur every
time an eventfd is armed with a read() I/O request on top of io_uring.

A significant performance improvement was made to Runtime.yield():
should the origin and destination worker thread targets of a yield()
call be the same, and should there be no pending tasks on said worker
thread, then the yield request will no-op. This cuts down the amount of
message passing down by over 47x, which reduces CPU usage significantly
and allows for more time to be dedicated to handling I/O.

The graceful shutdown strategy for Server was rewritten to prevent the
possibility of there being a deadlock waiting for coroutines bound to a
Server to clean their resources up. Upon the termination of a
Server.Connection's read loop, should the Server.Connection's write loop
be suspended, a boolean flag would be set and the write loop would be
forcefully woken up so that it may clean itself up. The same goes the
other way for a Server.Connection's write loop terminating and
forcefully waking up its possibly-suspended read loop.

The write loop for a Server.Connection was rewritten to not take a
linked list of nodes that comprise of messages that are pending to be
written, but instead to take a resizable FIFO ring buffer of bytes that
are pending to be written that should be flushed.

A lot of benchmarking was done to decide on how task queues for each
Worker should be managed. Four different strategies were benchmarked.
The benchmark is to send 100 million 10-byte ping messages on localhost.
The four strategies benchmarked are: N MPSC LIFO stacks per worker
thread, N MPSC FIFO queues per worker thread, 1 MPSC LIFO stack per
worker thread, and 1 MPSC FIFO queue per worker thread where N is the
number of worker threads instantiated on startup.

N MPSC LIFO stacks per worker thread: 29.288 seconds.
N MPSC FIFO queues per worker thread: 30.167 seconds.
1 MPSC LIFO stack per worker thread:  34.294 seconds.
1 MPSC FIFO queue per worker thread:  39.641 seconds.

As a result of these benchmarks, there are now N MPSC LIFO stacks per
worker thread. My perspective as to why the MPSC LIFO stack is the
fastest (for more context, the MPSC LIFO stack implementation resembles
a Treiber stack implementation) is because popping a message only
requires a single atomic XCHG operation. By reducing contention on the
producer end by having an N-length array of these stacks, there should
be little to no cross-thread contention at all in comparison to the MPSC
FIFO queue implementation.

On every worker loop iteration, all N MPSC LIFO stacks have their entire
contents popped and have their contained async frames resumed.

Benchmarking diagnostics that detail the number of events processed by
each worker, and the number of ping messages sent by the benchmark are
printed out every single second.

Pointer stability when ping messages are constructed in the main
entrypoint was fixed by re-ordering statements so that pointers are only
taken after the entirety of a single ping message is constructed.

To end off with some rough numbers of the performance of the system so
far, on my laptop, roughly 3.4 to 4 million 10-byte ping messages are
able to be sent per second. The memory usage in amidst the benchmark is
only roughly 0.3% of my 15.6GiB of RAM.