runtime: async mutex, simplify yield, store worker in thread-local var
Ported the word_lock implementation from parking_lot into Zig for
protecting shared state. Goals are to move away from yielding to and
from worker threads for protecting shared state, and to instead
protect shared state with mutices that take advantage of message
passing. The reason for choosing parking_lot's word_lock implementation
is because it optimizes against the fact that scheduling a suspended
thread to wake up is expensive, which applies to Rheia's
Runtime.schedule() function.

Cut down the fast-path condition in Runtime.yield() to just checking if
the origin and destination worker threads are the same, rather than if
the destination worker threads' task queue is empty. This reduces
contention on the task queues, and is not an important condition to
check for since, in the most cases, the worker threads' task queue is
empty whenever the origin and destination worker threads are the same.

Have Worker's store a pointer to themselves in a thread-local variable.
The current Worker instance may then be accessed via. a call to
Worker.getCurrent(). All Runtime functions that used to take an origin
worker thread ID parameter now longer require an origin worker thread ID
parameter to be passed in.

One idea to take into consideration in subsequent commits is to have the
Runtime be a globally-accessible variable rather than something that is
locally initalized and passed around as a pointer.

Reuse buffers in Server.runReadLoop() to reduce the amount of memory
allocations necessary.