rheia: intrusive data structures, mutex locks, and runtime api cleanup
Implemented a singly linked list, double-ended singly linked list, and
doubly-linked list that allows for custom Node types to be specified,
and allows for custom list pointer field names to be specified. All
operations that remove a node from a list nullifies the nodes' fields
such that the node may safely be reused in subsequent list operations.

Client/Server were heavily refactored to make use of the new Lock
primitive which takes advantage of the message passing system available
in Rheia. The rationale for this refactor is that the old strategy of
suspending a task, sending a message to resume the task on a thread
which contains state that is desired to be mutated, suspending the task
yet again, and sending a message to resume the task on the task's
original thread requires 2 messages to be passed unlike Lock requiring
solely just one message to be passed.

Making heavy use of the Lock primitive in Client/Server though unveils
that the overhead of the atomic operations underlying Lock is
non-negligible.

In spite of this, the use of shared state protected by the Lock primitive
has been decided for all forthcoming design decision in Rheia.

The Runtime API now introduces a separate scheduleTo(to, task) method
which schedules a task from the callers thread into the target thread.
On the other hand, the schedule(to, task) method's parameter set has
been changed to schedule(task). Runtime.schedule(task) now simply defers
the execution of the provided task on the same thread.

Given the decision to make use of shared state, a number of calls to
Runtime.yield() was removed, and suspended tasks now take in a
'worker_id' parameter so that functions handling these suspended tasks
know which thread to delegate the resumption of a task to.

Work still needs to be done to complete the heavy refactor of Client. A
lot of cleanup can be done of its API, and the use of timers in Client
needs to be refactored so that timers are properly cancelled upon
receiving a request to gracefully shutdown.

Yield calls in the main entrypoint benchmark are removed as the Lock
primitive provides sufficient yield points throughout the system.