rheia: global runtime, i/o cancellation, {m,s}p{m,s}c ring buffer
In this commit, Rheia's runtime has now all been consolidated into a
single runtime.zig module so that it is easy for the runtime to be
copied and pasted into other projects.

Loop has now been merged directly into Worker. In the future when
fallback I/O polling methods are to be implemented, all that would need
to be modified is the Worker as a whole rather than both a Worker and
Loop interface definition.

Rheia's runtime is now a global variable that is accessed by all other
components in Rheia. This cleans up a lot of the code as there is now no
longer a need to pass a reference to Runtime or Worker or Loop across
all modules.

All io_uring methods now take in a *runtime.Request. A runtime.Request
may be used as a handle to request for cancellation of any pending or
unwanted ongoing I/O requests via. runtime.cancel(*Request). With
cancellation now available, io_uring-based cancellable timers have now
been properly implemented and tested.

All sync-related primitives in Rheia have now been consolidated into a
single sync.zig.

An overall cleanup of both Client and Server has been done now that
runtime-related methods are exposed in the global scope.

A cleanup of io_uring methods was done, and error handling for all
io_uring-driven syscalls has been split into separate functions kept in
os.zig.

A lock-free ring buffer that supports single producers, multiple
producers, single consumers, and multiple consumers has been ported from
NetBSD. It is planned in subsequent commits to have the ring buffer
support bulk operations as well.

io_uring-specific benchmarks were written and are available in the
benchmarks/ folder. They were written to re-consider gauging the design
decision on whether or not we want to protect data shared across threads
using a mutex, or to encourage for message-passing instead, and whether
or not it is more performant for Rheia's TCP client pool to buffer
messages in userspace or to batch-submit send() requests to io_uring.

The results of the benchmarks are that 1) buffering messages in
userspace yields significantly more throughput, 2) we should not assume
io_uring simplifies how we fundamentally execute syscalls and handle
errors or results from syscalls but solely acts as an asynchronous
sycall interface, and 3) there is a linear increase in throughput the
more threads you introduce up to the total number of cores within your
system.

For the last point, on my system, I am able to submit roughly 200k
send() syscalls via. io_uring per second with a single thread. My laptop
in total has 4 cores. By having a thread-per-core model s.t. I have 4
threads spamming send() syscalls, I am able to submit roughly 800k
send() syscalls per second. If I add 1 or 2 more threads, I am able to
submit roughly about 850k to 950k send() syscalls per second.

As for the point about re-evaluating whether or not we want to protect
data shared across threads using a mutex, or to simply have all shared
data live within a single thread s.t. synchronization primitives may be
completedly removed, more benchmarks need to be done with message
passing. Hence why research is being done on my end on well-known
message passing primitives s.t. I ported NetBSD's lock-free ring buffer.

What caused the need to re-evaluate this design decision comes from the
fact that the introduction of sync.Lock in prior commits caused CPU
usage to spike from 70% per thread to 100% per thread. Prior to the
introduction of sync.Lock, roughly the same message throughput in
Rheia's benchmarks was yielded, and there was much less CPU usage which
implicates that the overhead of utilizing synchronization primitives is
not to be taken to be negligible.