rheia: remove use of locks, write more benchmarks and optimizations
In this commit, the throughput of a single Client has improved from 3.4
to 4 million messages per second my laptop to 8.5 to 9 million messages
per second. This was accomplished by having the responsibility of
buffering messages be done from the callers end rather than the writers
end. The rationale for the improvements is that on a single thread,
roughly 24 million 10-byte messages can be allocated per second. Given
that message serialization is very fast, there would be no benefit to
parallelizing the buffering of messages by having buffering done on the
writesr end. Hence why it makes much more sense to have the caller
perform buffering on their end instead.

Use of locks were removed in favor of message passing. The rationale is
that message passing naturally allows for far less CPU cycles to be
expended, and thus strains the CPU much less in comparison to locks. CPU
usage per thread has now been reduced from an average of 100% when
running benchmarks down to 0-5%.

Benchmarks were written for testing the performance of Zig's async
frames in tandem with io_uring. In comparison to Go which makes use of
epoll, there is a noticeable 5-20% drop in throughput. Nonetheless, as
io_uring still has many more improvements that have yet to be
introduced, io_uring will prevail as the basis for all worker event
loops in Rheia.

The number of workers determined to be spawned on startup was
incorrectly computed, causing single-threaded builds of Rheia to not
work. The calculation has been fixed in this commit.

Added single-producer and fixed the multi-producer variant for the
lock-free ring buffer that may be founded in mpmc.zig. Tests have been
fixed as well.