rheia: rpc, crypto, cancellation, tx verifier, cpu-bound thread pool
Request/response has been fully implemented both on the client side and
server side.

Synchronization primitives that are intended to be used in a
single-threaded context have been implemented. Primitives akin to a
mutex (sync.Mutex), condition variable (sync.Parker), and a semaphore or
wait group (sync.WaitGroup) have been implemented.

Rewrote all async cancellation logic to make use of a Context. A Context
is a linked list of callbacks that is called when Context.cancel() is
called. Context's are not intended to be used across multiple threads.
Rheia's Context implementation is essentially a simplified version of
Go's context.Context, and folly's CancellationToken.

All functions that may possibly block now take in a Context, and
register appropriate callbacks to the passed-in Context such that the
function may be cancellable.

Based on observed benchmarks, both Client and Server were entirely
re-written to now support cancellation via. Context, and to now operate
entirely single-threaded (for the time being). Once the majority of
Rheia's features are implemented, work will be done to lift the Client
and Server implementations to utilize a thread-per-core architecture. It
is important to note however that for Rheia, the workload is generally
CPU-bound as the majority of CPU usage is spent on transaction signature
verification rather. Hence, it is unlikely that a thread-per-core
architecture will highly benefit Rheia's performance.

All runtime functions that submit a request to perform a syscall to an
io_uring file descriptor now take in a Context. All runtime methods that
may potentially block therefore now support cancellation.

A transaction verifier has been implemented which batches up incoming
received transactions from the servers end into batches of 64
transactions, and schedules them to have their signatures verified into
a CPU-bound thread pool. The CPU-bound thread pool implementation comes
from kprotty/zap, and is available at Pool.zig. The transaction verifier
has a debouncing mechanism where if transactions stop coming in, and
there is less than 64 transactions queued into the verifier, then after
100 milliseconds the remaining transactions will be scheduled to be
verified in Rheia's CPU-bound thread pool.

CPU-bound operations that are meant to be executed in the CPU-bound
thread pool can be initiated by making use of the functions
runtime.startCpuBoundOperation() and runtime.endCpuBoundOperation().
The latter function reschedules the callers frame to the main thread of
Rheia.

Based on benchmarks performed locally on my laptop, a total of roughly
130k transactions can be received and have their transactions verified
per second.