art: fix node16 child insertion order, reduce internal node size
Spent the day debugging the structure of Rheia's adaptive radix tree.

Added convenience methods for printing interior/leaf nodes.

Reduced `partial_len` down from a u16 to a u8 as the max partial length
for an interior node is 10.

After spending some time inserting a bunch of keys and visualizing the
tree structure, spotted and fixed a problem to do with the order in
which nodes with 16 children append new children to themselves.

Updated travisstaloch/art.zig in the mempool benchmarks to the latest
version.