hash_map: fix collision check bugs on get/delete
After injecting artificial hashes where the 64 bits are the same as
suggested by Paul Khuong, a bug became apparent in get/delete: entries
that were previously shifted forward in the bucket they are assigned to
upon insertion would never be linearly probed.

This apparently happened due to get/delete not linearly probing through
the bucket that a provided key is mapped to. The fix is to simply probe
the bucket.

The test Paul suggested has also been provided as a unit test to
hash_map.zig.