hash_map: use avx2 instructions to compare keys
The key comparison routine in the hash map has had its branches
optimized against inputs that are cryptographic hashes by assuming that
if the first 64 bits of cryptographic hashes 'a' and 'b' are equivalent,
then 'a' and 'b' are most likely equivalent.

This assumption holds with high likelihood given that cryptographic
hashes naturally are collision-resistant under the random oracle model.