chain: cache last N finalized blocks, respond to block pull requests
A ring buffer was written to cache the last N finalized blocks. It is
assumed that any reference counting happens prior to pushing a block
into a ring buffer, and after evicting a block from the ring buffer.

Logic for handling block pull requests has been implemented. In the case
the latest known block height + 1 is requested to be pulled, the
most-preferred block proposal cached in the nodes' sampler (if any) is
provided to the requester. In the case the last N finalized blocks is
requested to be pulled, the block is provided to the requester from the
nodes' block ring buffer. Any other block pull request with some other
requested block height will yield a null block to the requester.

Code has been written such that a Node now owns and manages multiple TCP
client pool instances, and stores them in a hash map keyed by their
individual destination addresses. The transaction sending benchmark in
the main entrypoint of Rheia has been updated to reflect the fact that a
Node now owns a set of net.Client instances.