chain: cache valid block proposal responses from peers
In order to reduce network load, hint to peers the previous block
response that was received so that they do not need to redundantly send
the same block response yet again on the next round trip.

Fixed lru.HashMap.clear() not resetting head and tail pointers.

Moved all the owned references to Block's to both the block proposal
cache and block sampler within the consensus protocol loop.

Fix not redundantly sending the same block response yet again should a
cache hint sent in a request which triggers a cache hit.

Fix block response packet length in the case of a cache hit.

Reset exponential backoff delay in the case a node manages to connect to
and interact with at least one peer to perform consensus.