node: implement push side of gossip protocol
Implemented the 'push' side of Rheia's gossip protocol. When nodes
receive a new, valid transaction that has passed signature verification,
the transaction will be queued to be pushed to neighboring peers.

Transactions that have been pushed to a peer will not be pushed to the
same peer. This guarantee only roughly holds as the set of all
transactions that have been pushed to a peer is kept track of using a
large global LRU cache.

Implemented a getOrPut() API to the robin hood hash table. The API is
used in order to ignore incoming transactions that a node is already
aware of, and to acknowledge incoming transactions that a node is not
yet aware of.

Reduce packet size to 65536 (2^16) bytes. The size limitation will be
lifted once net.Client and net.Server has been adapted to allow for a
customizable message buffer limit. Otherwise, net.Client in particular
right now blocks a caller due to backpressure when a caller attempts to
buffer more than 65536 bytes to be sent at a time (65536 in this case
being a hardcoded limit as of now).

Removed unnecessary `await async` prefixes when calling async functions
in the main entrypoint.

Print out 'shutting down...' logs and ensure all deinit() methods are
not able to return errors to make cleanup of the main entrypoint clean.