runtime: fix race condition in rearming eventfd
Previously, should an attempt to arm a read() request to an eventfd
fail, the event loop would block indefinitely should no I/O completions
be made. The expected behavior should be that the event loop should not
block, should continue processing any new I/O completions, and should
re-attempt to re-arm a read() request to its eventfd.

This glaring detail is most likely the cause of the following deadlock:

A graceful shutdown is requested on a node. The TCP listener of the node
is shut down, then shutdown(SHUT_RD) is called on all incoming clients to
the TCP listener. All of the incoming clients disconnect.

From the clients' perspective, having all of its pooled connections drop
would engage the client to attempt to reconnect to the listener.

Once the TCP listener of the node is shut down though, the client is
then requested to attempt to perform a graceful shutdown.

The client first cancels all live tasks, then suspends the current
coroutine until all of its pooled connections have cleaned themselves
up.

A deadlock then emerges: the current coroutine is deadlocked waiting for
the clients' pooled connections to have cleaned themselves up.

After doing lots of printf-debugging, the deadlock was due to pooled
connections being deadlocked on a sync.Mutex that is responsible for
only allowing one connection to attempt to perform a connect() request
at any moment in time.

More specifically, connections deadlocked on a sync.Mutex were not being
resumed after a prior connect() request had completed (with or without
error).

It appears that sync.Mutex was doing its job however. On
sync.Mutex.release(), the next waiter in FIFO order would be popped and
queued into the runtime to be resumed.

However, after opening up gdb, it appears that the next waiter was
popped, though apparently not queued as a 'pending task' with regards to
the runtime.

There didn't appear to be any glaring holes in the sync.Mutex logic, and
so after an investigation into Runtime, one potential glaring hole which
could potentially cause a deadlock was the cross-thread event
notification system which utilizes an eventfd.

And as it turns out, rewriting how the eventfd is rearmed appears to
have caused the deadlock to not longer prevail. There is a possibility
that I might have just lowered the likelihood in which the deadlock
prevails, though I have attempted to gracefully shutdown a node on my
laptop about 1000 times now and such a deadlock has not happened at
least once. In prior tests, such a deadlock happened once every 5 to 10
attempts to gracefully shutdown a node.

The condition for a net.Client to spawn and pool a new connection has
been made stringent s.t. a new pooled connection may only be spawned
should there 1) be no connections in the pool, or 2) the last connect()
attempt had not failed and the pool is not full.